Delete data files after upload finishes

+++ src/snowflake/connector/storage_client.py
@@ -409,8 +409,14 @@ class SnowflakeStorageClient(ABC):
     def _abort_multipart_upload(self) -> None:
         pass
 
-    def __del__(self):
-        logger.debug(f"cleaning up tmp dir: {self.tmp_dir}")
-        shutil.rmtree(self.tmp_dir)
+    def delete_client_data(self):
+        """Deletes the tmp_dir and closes the source stream belonging to this client.
+        This function is idempotent."""
+        if os.path.exists(self.tmp_dir):
+            logger.debug(f"cleaning up tmp dir: {self.tmp_dir}")
+            shutil.rmtree(self.tmp_dir)
         if self.meta.real_src_stream and not self.meta.real_src_stream.closed:
             self.meta.real_src_stream.close()
+
+    def __del__(self):
+        self.delete_client_data()
